Phase 1: 轻量化重构 — 去除 Docker/PostgreSQL/Redis，改用 SQLite + 内存缓存
- 删除 7 个 Docker 文件 + redis_client.py
- 数据库: PostgreSQL → SQLite (aiosqlite)
- 缓存: Redis → 内存 TTL dict + asyncio.Lock
- 模型: BigInteger → Integer (SQLite 自增兼容)
- Schema: fund_name/buy_date 改为可选
- JSON: 全局自定义编码器 Decimal→float
- 新增 start.bat / start.sh 一键启动脚本

+++ backend/app/models/data_source.py
@@ -6,7 +6,7 @@ Tracks external data fetches for audit and observability.
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Integer, String, Text, func
+from sqlalchemy import DateTime, Integer, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -18,7 +18,7 @@ class DataSourceLog(Base):
     __tablename__ = "data_source_logs"
 
     id: Mapped[int] = mapped_column(
-        BigInteger, primary_key=True, autoincrement=True
+        Integer, primary_key=True, autoincrement=True
     )
     source_name: Mapped[str | None] = mapped_column(
         String(50), default=None, comment="Name of the data source"
